requirements.txt
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.2.1
filelock==3.8.0
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pipenv==2022.9.24
platformdirs==2.5.2
pygame==2.1.2
pymongo==4.3.3
python-dotenv==0.21.1
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.2.2
